Add buffer for nmi_occurred flag
The buffer is now set one cycle sooner and flushed to the primary
latch on the following cycle. This effectively causes the NMI
suppression behavior related to PPUSTATUS read to be "shifted back"
by 1 cycle, i.e. it now affects cycles 0-2 of the scanline instead
of 1-3.

This allows all tests in ppu_vbl_nmi to pass and also fixes a
regression affecting the Tengen version of Ms. Pac-Man.